requirements.txt
django 
joblib
pandas
torch
scikit-learn
channels
websocket
websocket-client
daphne
flask 
requests
whitenoise
scipy


Django>=4.2
torch
torchvision
librosa
matplotlib
numpy
torchaudio 
tensorflow
opencv-python
pillow
lime